Prefer using http Host header for client-side signing
If the outgoing http request has specified the http Host header, that
host should be used to formulate the baseString that is in turned used
to create the signature. If it is not specified, fall back to the host
in the request URL.

This is required to properly sign oauth requests that travel through a
proxy.